Update/fix Contributing Section
Reword the Contributing section to be more focused on the two important aspects: how to contribute and the community guidelines.  Also move it above `Issues` section on the page for better visibility.

This also "fixes" the section by removing a broken link and linking to the README in the `contributing` folder for better readability.